+++ imod/tests/test_mf6/test_mf6_ssm.py
@@ -51,6 +51,29 @@ def test_transport_model_multi_bcs_rendering(flow_model_with_concentration):
     assert actual == expected
 
 
+@pytest.mark.usefixtures("flow_model_with_concentration")
+def test_transport_model_rendering_settings(flow_model_with_concentration):
+    directory = pathlib.Path("mymodel")
+    globaltimes = np.array(["2000-01-01"], dtype="datetime64[ns]")
+    m = SourceSinkMixing.from_flow_model(
+        flow_model_with_concentration, "salinity", save_flows=True, print_flows=True
+    )
+    actual = m.render(directory, "river", globaltimes, True)
+    expected = textwrap.dedent(
+        """\
+        begin options
+          print_flows
+          save_flows
+        end options
+
+        begin sources
+          riv-1 AUX salinity
+        end sources
+        """
+    )
+    assert actual == expected
+
+
 def test_wrong_dtype():
     SourceSinkMixing(
         np.array(["a", "b"]),